@@ -6,12 +6,12 @@ build-backend = "hatchling.build"
 name = "infisical"
 description = 'Official Infisical SDK for Python'
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 license = "MIT"
 authors = []
 maintainers = [
   { name = "Yohann MARTIN", email = "contact@codexus.fr" },
-  { name = "Tony Dang", email = "tony@infisical.com"}
+  { name = "Tony Dang", email = "tony@infisical.com" },
 ]
 classifiers = [
   "Development Status :: 5 - Production/Stable",
@@ -24,7 +24,6 @@ classifiers = [
   "Intended Audience :: Developers",
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python",
-  "Programming Language :: Python :: 3.7",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
@@ -33,10 +32,10 @@ classifiers = [
   "Programming Language :: Python :: Implementation :: PyPy",
 ]
 dependencies = [
-  "requests ==2.31.0",
-  "pydantic >=1.6.2,!=1.7,!=1.7.1,!=1.7.2,!=1.7.3,!=1.8,!=1.8.1,<2.0.0",
-  "pycryptodomex >=3.17,<4.0.0",
-  "pynacl >=1.5.0,<2.0.0"
+  "requests==2.31.0",
+  "pydantic>=1.7.4,!=1.8,!=1.8.1,!=2.0.0,!=2.0.1,!=2.1.0,<3.0.0",
+  "pycryptodomex>=3.17,<4.0.0",
+  "pynacl>=1.5.0,<2.0.0",
 ]
 dynamic = ["version"]
 
@@ -47,68 +46,55 @@ Source = "https://github.com/Infisical/infisical-python"
 
 [project.optional-dependencies]
 test = [
-  "pytest >=7.1.3,<8.0.0",
-  "coverage[toml] >= 6.5.0,< 8.0",
-  "responses ==0.23.1"
+  "pytest>=7.1.3,<8.0.0",
+  "coverage[toml]>=6.5.0,< 8.0",
+  "responses==0.23.3",
+  "python-dotenv==1.0.0"
 ]
 dev = [
-  "mypy ==1.1.1",
-  "ruff ==0.0.261",
-  "black ==23.3.0",
-  "isort >=5.0.6,<6.0.0",
-  "devtools[pygments] ==0.11.0",
+  "mypy==1.5.1",
+  "ruff==0.0.292",
+  "black==23.9.1",
+  "devtools[pygments]==0.12.2",
 
-  "types-requests ==2.28.11.17"
+  "types-requests==2.31.0.7",
 ]
 
 [tool.hatch.version]
 path = "infisical/__version__.py"
 
 [tool.hatch.build.targets.sdist]
-exclude = [
-  "/.github",
-  "/.vscode",
-]
-
-[tool.isort]
-profile = "black"
-known_third_party = ["infisical", "pydantic", "Cryptodome", "nacl", "responses"]
+exclude = ["/.github", "/.vscode"]
 
 [tool.mypy]
 strict = true
+plugins = ["pydantic.mypy"]
 
 [tool.coverage.run]
 parallel = true
-source = [
-    "tests",
-    "infisical"
-]
+source = ["tests", "infisical"]
 
 [tool.coverage.report]
-exclude_lines = [
-  "no cov",
-  "if __name__ == .__main__.:",
-  "if TYPE_CHECKING:",
-]
+exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
 
 [tool.ruff]
-target-version = "py37"
+target-version = "py38"
 select = [
-  "E",  # pycodestyle errors
-  "W",  # pycodestyle warnings
-  "F",  # pyflakes
-  # "I",  # isort
-  "C",  # flake8-comprehensions
-  "B",  # flake8-bugbear
+  "E", # pycodestyle errors
+  "W", # pycodestyle warnings
+  "F", # pyflakes
+  "I", # isort
+  "C", # flake8-comprehensions
+  "B", # flake8-bugbear
 ]
 ignore = [
-  "E501",  # line too long, handled by black
-  "B008",  # do not perform function calls in argument defaults
-  "C901",  # too complex
+  "E501", # line too long, handled by black
+  "B008", # do not perform function calls in argument defaults
+  "C901", # too complex
 ]
 
 [tool.ruff.per-file-ignores]
 "__init__.py" = ["F401"]
 
 [tool.ruff.isort]
-known-third-party = ["infisical"]
+known-third-party = ["infisical", "pydantic", "Cryptodome", "nacl", "responses"]
