@@ -1,6 +1,6 @@
 import os
 from datetime import datetime, timedelta
-from typing import TYPE_CHECKING, Union
+from typing import TYPE_CHECKING, List, Union
 
 from typing_extensions import Literal
 
@@ -12,7 +12,13 @@ from infisical.models.models import SecretBundle
 from infisical.services.secret_service import SecretService
 
 
-def get_all_secrets_helper(instance: "InfisicalClient", environment: str, path: str, include_imports: bool, attach_to_os_environ: bool):
+def get_all_secrets_helper(
+    instance: "InfisicalClient",
+    environment: str,
+    path: str,
+    include_imports: bool,
+    attach_to_os_environ: bool,
+) -> List[SecretBundle]:
     try:
         if not instance.client_config:
             raise Exception("Failed to find client config")
@@ -31,14 +37,14 @@ def get_all_secrets_helper(instance: "InfisicalClient", environment: str, path: 
             environment=environment,
             path=path,
             workspace_key=instance.client_config.workspace_config.workspace_key,
-            include_imports=include_imports
+            include_imports=include_imports,
         )
 
         for secret_bundle in secret_bundles:
             cache_key = f"{secret_bundle.type}-{secret_bundle.secret_name}"
             instance.cache[cache_key] = secret_bundle
             if attach_to_os_environ:
-                os.environ[secret_bundle.secret_name] = secret_bundle.secret_value
+                os.environ[secret_bundle.secret_name] = secret_bundle.secret_value or ""
 
         return secret_bundles
     except Exception as exc:
@@ -54,7 +60,7 @@ def get_secret_helper(
     type: Literal["shared", "personal"],
     environment: str,
     path: str,
-):
+) -> SecretBundle:
     cache_key = f"{type}-{secret_name}"
     cached_secret: Union[SecretBundle, None] = None
     try:
@@ -117,7 +123,7 @@ def create_secret_helper(
     type: Literal["shared", "personal"],
     environment: str,
     path: str,
-):
+) -> SecretBundle:
     try:
         if not instance.client_config:
             raise Exception("Failed to find client config")
@@ -159,7 +165,7 @@ def update_secret_helper(
     type: Literal["shared", "personal"],
     environment: str,
     path: str,
-):
+) -> SecretBundle:
     try:
         if not instance.client_config:
             raise Exception("Failed to find client config")
@@ -200,7 +206,7 @@ def delete_secret_helper(
     type: Literal["shared", "personal"],
     environment: str,
     path: str,
-):
+) -> SecretBundle:
     try:
         if not instance.client_config:
             raise Exception("Failed to find client config")
